requirements.txt
pandas
fastapi
uvicorn
pydantic
numpy
mysql-connector